last pushes

+++ src/train.py
@@ -331,7 +331,7 @@ class Trainer():
         for self.batch_number, batch_data in enumerate(self.training_generator):
             input, transcription = batch_data
 
-            input, transcription = input.float().to(self.device), transcription.long().to(self.device)
+            input, transcription = input.to(self.device), transcription.to(self.device)
                       
             if self.batch_number == 0: logger.info(f"input.size(): {input.size()}")
